feat: drawer menu comp and setting menu
