+++ client/src/components/Typography/TextLabel.js
@@ -1,25 +1,8 @@
 import { Typography } from "antd";
 import styled from "styled-components";
-import { mq, theme } from "constants/theme";
+import { theme } from "constants/theme";
 
 const { Text } = Typography;
-
-/*  Reference from theme.js
-    xxsmall: "1rem",
-    xsmall: "1.1rem",
-    small: "1.2rem",
-    medium: "1.4rem",
-    large: "1.6rem",
-    xlarge: "1.8rem"
-
-    usage:
-    <TextLabel
-        block="true"
-        color={DARK_GRAY}
-        size={theme.typography.size.medium}
-    >
-
-*/
 
 const TextLabel = styled(Text)`
   &.ant-typography {
